+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog for NeoFS Node
 - NPE in metabase V3->4 migration routine (#3212)
 - "new transaction callback finished with error" logs for "already exists in mempool" error (#3218)
 - Panic if logger is not set in FSTree (#3220)
+- Not initialized FSTree in the write-cache (#3220)
 
 ### Changed